refine the code according review advice test=develop
