+++ python/paddle/fluid/tests/unittests/test_fill_any_like_op.py
@@ -97,195 +97,5 @@ class TestFillAnyLikeOpFloat16(TestFillAnyLikeOp):
         self.dtype = np.float16
 
 
-class TestFillAnyLikeOp_attr_out(unittest.TestCase):
-    """ Test fill_any_like op(whose API is full_like) for attr out. """
-
-    def test_attr_tensor_API(self):
-        startup_program = fluid.Program()
-        train_program = fluid.Program()
-        with fluid.program_guard(train_program, startup_program):
-            fill_value = 2.0
-            input = fluid.data(name='input', dtype='float32', shape=[2, 3])
-            output = paddle.full_like(input, fill_value)
-            output_dtype = paddle.full_like(input, fill_value, dtype='float32')
-
-            place = fluid.CPUPlace()
-            if fluid.core.is_compiled_with_cuda():
-                place = fluid.CUDAPlace(0)
-            exe = fluid.Executor(place)
-            exe.run(startup_program)
-
-            img = np.array([[1, 2, 3], [4, 5, 6]]).astype(np.float32)
-
-            res = exe.run(train_program,
-                          feed={'input': img},
-                          fetch_list=[output])
-
-            out_np = np.array(res[0])
-            self.assertTrue(
-                not (out_np - np.full_like(img, fill_value)).any(),
-                msg="full_like output is wrong, out = " + str(out_np))
-
-
-class TestFillAnyLikeOpError(unittest.TestCase):
-    def test_errors(self):
-        with program_guard(Program(), Program()):
-            #for ci coverage
-
-            input_data = fluid.data(name='input', dtype='float32', shape=[2, 3])
-            output = paddle.full_like(input_data, 2.0)
-
-            def test_input_dtype():
-                paddle.full_like
-
-            self.assertRaises(
-                ValueError,
-                paddle.full_like,
-                x=input_data,
-                fill_value=2,
-                dtype='uint4')
-            self.assertRaises(
-                TypeError,
-                paddle.full_like,
-                input=input_data,
-                fill_value=2,
-                dtype='int16')
-
-
-class ApiOnesLikeTest(unittest.TestCase):
-    def test_out(self):
-        with fluid.program_guard(fluid.Program()):
-            data = fluid.data(shape=[10], dtype="float64", name="data")
-            ones = paddle.ones_like(data, device="cpu")
-            place = fluid.CPUPlace()
-            exe = fluid.Executor(place)
-            result, = exe.run(feed={"data": np.random.rand(10)},
-                              fetch_list=[ones])
-            expected_result = np.ones(10, dtype="float64")
-        self.assertEqual((result == expected_result).all(), True)
-
-        with fluid.program_guard(fluid.Program()):
-            data = fluid.data(shape=[10], dtype="float64", name="data")
-            ones = paddle.ones_like(data, device="cpu", dtype="float32")
-            place = fluid.CPUPlace()
-            exe = fluid.Executor(place)
-            result, = exe.run(feed={"data": np.random.rand(10)},
-                              fetch_list=[ones])
-            expected_result = np.ones(10, dtype="float32")
-        self.assertEqual((result == expected_result).all(), True)
-
-        with fluid.program_guard(fluid.Program()):
-            data = fluid.data(shape=[10], dtype="float64", name="data")
-            ones = paddle.ones_like(data)
-            place = fluid.CPUPlace()
-            exe = fluid.Executor(place)
-            result, = exe.run(feed={"data": np.random.rand(10)},
-                              fetch_list=[ones])
-            expected_result = np.ones(10, dtype="float32")
-        self.assertEqual((result == expected_result).all(), True)
-
-
-class ApiZerosLikeTest(unittest.TestCase):
-    def test_out(self):
-        with fluid.program_guard(fluid.Program()):
-            data = fluid.data(shape=[10], dtype="float64", name="data")
-            zeros = paddle.zeros_like(data, device="cpu")
-            place = fluid.CPUPlace()
-            exe = fluid.Executor(place)
-            result, = exe.run(feed={"data": np.random.rand(10)},
-                              fetch_list=[zeros])
-            expected_result = np.zeros(10, dtype="float64")
-        self.assertEqual((result == expected_result).all(), True)
-
-        with fluid.program_guard(fluid.Program()):
-            data = fluid.data(shape=[10], dtype="float64", name="data")
-            zeros = paddle.zeros_like(data, device="cpu", dtype="float32")
-            place = fluid.CPUPlace()
-            exe = fluid.Executor(place)
-            result, = exe.run(feed={"data": np.random.rand(10)},
-                              fetch_list=[zeros])
-            expected_result = np.zeros(10, dtype="float32")
-        self.assertEqual((result == expected_result).all(), True)
-
-        with fluid.program_guard(fluid.Program()):
-            data = fluid.data(shape=[10], dtype="float64", name="data")
-            zeros = paddle.zeros_like(data)
-            place = fluid.CPUPlace()
-            exe = fluid.Executor(place)
-            result, = exe.run(feed={"data": np.random.rand(10)},
-                              fetch_list=[zeros])
-            expected_result = np.zeros(10, dtype="float32")
-        self.assertEqual((result == expected_result).all(), True)
-
-
-class TestOnesZerosError(unittest.TestCase):
-    def test_errors(self):
-        def test_device_error1():
-            with fluid.program_guard(fluid.Program(), fluid.Program()):
-                data = fluid.data(name="data", shape=[10], dtype="float32")
-                paddle.ones_like(data, device="opu")
-
-        self.assertRaises(ValueError, test_device_error1)
-
-        def test_device_error2():
-            with fluid.program_guard(fluid.Program(), fluid.Program()):
-                data = fluid.data(name="data", shape=[10], dtype="float32")
-                paddle.ones_like(data, dtype="float")
-
-        self.assertRaises(ValueError, test_device_error2)
-
-        def test_device_error3():
-            with fluid.program_guard(fluid.Program(), fluid.Program()):
-                data = fluid.data(name="data", shape=[10], dtype="float32")
-                paddle.zeros_like(data, device="opu")
-
-        self.assertRaises(ValueError, test_device_error3)
-
-        def test_device_error4():
-            with fluid.program_guard(fluid.Program(), fluid.Program()):
-                data = fluid.data(name="data", shape=[10], dtype="float32")
-                paddle.zeros_like(data, dtype="float")
-
-        self.assertRaises(ValueError, test_device_error4)
-
-        def test_ones_like_type_error():
-            with fluid.program_guard(fluid.Program(), fluid.Program()):
-                fluid.layers.ones_like([10], dtype="float")
-
-        self.assertRaises(TypeError, test_ones_like_type_error)
-
-        def test_ones_like_dtype_error():
-            with fluid.program_guard(fluid.Program(), fluid.Program()):
-                data = fluid.data(name="data", shape=[10], dtype="float16")
-                fluid.layers.ones_like(data, dtype="float32")
-
-        self.assertRaises(TypeError, test_ones_like_dtype_error)
-
-        def test_ones_like_out_type_error():
-            with fluid.program_guard(fluid.Program(), fluid.Program()):
-                data = fluid.data(name="data", shape=[10], dtype="float32")
-                fluid.layers.ones_like(data, dtype="float32", out=[10])
-
-        self.assertRaises(TypeError, test_ones_like_out_type_error)
-
-        def test_ones_like_out_dtype_error():
-            with fluid.program_guard(fluid.Program(), fluid.Program()):
-                data = fluid.data(name="data", shape=[10], dtype="float32")
-                out = fluid.data(name="out", shape=[10], dtype="float16")
-                fluid.layers.ones_like(data, dtype="float32", out=out)
-
-        self.assertRaises(TypeError, test_ones_like_out_dtype_error)
-
-
-class ImperativeModeTest(unittest.TestCase):
-    def test_full_like_imperative(self):
-        with paddle.imperative.guard():
-            input = paddle.arange(6, 10, dtype='float32')
-            out = paddle.full_like(input, fill_value=888.88, dtype='float32')
-            out_numpy = np.random.random((4)).astype("float32")
-            out_numpy.fill(888.88)
-            self.assertTrue((out.numpy() == out_numpy).all(), True)
-
-
 if __name__ == "__main__":
     unittest.main()
